refactor: group modules into domain packages
