+++ src/cli/__init__.py
@@ -1,0 +1,5 @@
+"""Command-line interface package."""
+
+from .main import main
+
+__all__ = ["main"]
