+++ src/engine/processing_engine.py
@@ -24,12 +24,12 @@ from io_utils.loader import (
     load_evolution_prompt,
     load_role_ids,
 )
-from model_factory import ModelFactory
+from io_utils.persistence import atomic_write, read_lines
+from io_utils.service_loader import load_services
 from models import ServiceInput
-from persistence import atomic_write, read_lines
+from models.factory import ModelFactory
 from runtime.environment import RuntimeEnv
-from service_loader import load_services
-from settings import Settings
+from runtime.settings import Settings
 from utils import ErrorHandler, LoggingErrorHandler
 
 # Helper functions migrated from cli for reuse.
