+++ src/runtime/__init__.py
@@ -2,5 +2,6 @@
 """Runtime package exposing the :class:`RuntimeEnv` singleton."""
 
 from .environment import RuntimeEnv
+from .settings import Settings, load_settings
 
-__all__ = ["RuntimeEnv"]
+__all__ = ["RuntimeEnv", "Settings", "load_settings"]
