+++ tests/test_cli_apply_args_to_settings.py
@@ -9,7 +9,7 @@ from types import SimpleNamespace
 
 import pytest
 
-import cli
+import cli.main as cli
 
 os.environ.setdefault("OPENAI_API_KEY", "test-key")
