+++ tests/test_cli_mapping_helpers.py
@@ -8,7 +8,7 @@ from types import SimpleNamespace
 
 import pytest
 
-import cli_mapping
+from cli import mapping as cli_mapping
 from core import mapping
 from models import Contribution, MappingSet, ServiceEvolution
