+++ tests/test_e2e_cli_mapping.py
@@ -7,7 +7,7 @@ import sys
 from pathlib import Path
 from types import SimpleNamespace
 
-import cli
+import cli.main as cli
 from core import mapping
 from models import Contribution, MappingSet
