+++ tests/test_loading.py
@@ -16,8 +16,8 @@ from io_utils.loader import (
     load_role_ids,
     load_roles,
 )
+from io_utils.service_loader import load_services
 from models import JobToBeDone
-from service_loader import load_services
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1] / "src"))
